Add Identity API

+++ net/enterprise.identity/src/Identity.API/Data/ApplicationDbContextSeed.cs
@@ -5,6 +5,8 @@ using System.IO.Compression;
 using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using Identity.API.Extensions;
+using Identity.API.Models;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Logging;
@@ -19,7 +21,7 @@ namespace Identity.API.Data
         public async Task SeedAsync(ApplicationDbContext context, IHostingEnvironment env,
             ILogger<ApplicationDbContextSeed> logger, IOptions<AppSettings> settings, int? retry = 0)
         {
-            int retryForAvaiability = retry.Value;
+            var retryForAvaiability = retry.Value;
 
             try
             {
@@ -36,10 +38,7 @@ namespace Identity.API.Data
                     await context.SaveChangesAsync();
                 }
 
-                if (useCustomizationData)
-                {
-                    GetPreconfiguredImages(contentRootPath, webroot, logger);
-                }
+                if (useCustomizationData) GetPreconfiguredImages(contentRootPath, webroot, logger);
             }
             catch (Exception ex)
             {
@@ -56,17 +55,15 @@ namespace Identity.API.Data
 
         private IEnumerable<ApplicationUser> GetUsersFromFile(string contentRootPath, ILogger logger)
         {
-            string csvFileUsers = Path.Combine(contentRootPath, "Setup", "Users.csv");
+            var csvFileUsers = Path.Combine(contentRootPath, "Setup", "Users.csv");
 
-            if (!File.Exists(csvFileUsers))
-            {
-                return GetDefaultUser();
-            }
+            if (!File.Exists(csvFileUsers)) return GetDefaultUser();
 
             string[] csvheaders;
             try
             {
-                string[] requiredHeaders = {
+                string[] requiredHeaders =
+                {
                     "cardholdername", "cardnumber", "cardtype", "city", "country",
                     "email", "expiration", "lastname", "name", "phonenumber",
                     "username", "zipcode", "state", "street", "securitynumber",
@@ -82,12 +79,16 @@ namespace Identity.API.Data
             }
 
             List<ApplicationUser> users = File.ReadAllLines(csvFileUsers)
-                        .Skip(1) // skip header column
-                        .Select(row => Regex.Split(row, ",(?=(?:[^\"]*\"[^\"]*\")*[^\"]*$)"))
-                        .SelectTry(column => CreateApplicationUser(column, csvheaders))
-                        .OnCaughtException(ex => { logger.LogError(ex.Message); return null; })
-                        .Where(x => x != null)
-                        .ToList();
+                .Skip(1) // skip header column
+                .Select(row => Regex.Split(row, ",(?=(?:[^\"]*\"[^\"]*\")*[^\"]*$)"))
+                .SelectTry(column => CreateApplicationUser(column, csvheaders))
+                .OnCaughtException(ex =>
+                {
+                    logger.LogError(ex.Message);
+                    return null;
+                })
+                .Where(x => x != null)
+                .ToList();
 
             return users;
         }
@@ -95,15 +96,12 @@ namespace Identity.API.Data
         private ApplicationUser CreateApplicationUser(string[] column, string[] headers)
         {
             if (column.Count() != headers.Count())
-            {
-                throw new Exception($"column count '{column.Count()}' not the same as headers count'{headers.Count()}'");
-            }
+                throw new Exception(
+                    $"column count '{column.Count()}' not the same as headers count'{headers.Count()}'");
 
-            string cardtypeString = column[Array.IndexOf(headers, "cardtype")].Trim('"').Trim();
-            if (!int.TryParse(cardtypeString, out int cardtype))
-            {
+            var cardtypeString = column[Array.IndexOf(headers, "cardtype")].Trim('"').Trim();
+            if (!int.TryParse(cardtypeString, out var cardtype))
                 throw new Exception($"cardtype='{cardtypeString}' is not a number");
-            }
 
             var user = new ApplicationUser
             {
@@ -126,7 +124,7 @@ namespace Identity.API.Data
                 NormalizedEmail = column[Array.IndexOf(headers, "normalizedemail")].Trim('"').Trim(),
                 NormalizedUserName = column[Array.IndexOf(headers, "normalizedusername")].Trim('"').Trim(),
                 SecurityStamp = Guid.NewGuid().ToString("D"),
-                PasswordHash = column[Array.IndexOf(headers, "password")].Trim('"').Trim(), // Note: This is the password
+                PasswordHash = column[Array.IndexOf(headers, "password")].Trim('"').Trim() // Note: This is the password
             };
 
             user.PasswordHash = _passwordHasher.HashPassword(user, user.PasswordHash);
@@ -137,89 +135,79 @@ namespace Identity.API.Data
         private IEnumerable<ApplicationUser> GetDefaultUser()
         {
             var user =
-            new ApplicationUser()
-            {
-                CardHolderName = "DemoUser",
-                CardNumber = "4012888888881881",
-                CardType = 1,
-                City = "Redmond",
-                Country = "U.S.",
-                Email = "demouser@microsoft.com",
-                Expiration = "12/20",
-                Id = Guid.NewGuid().ToString(),
-                LastName = "DemoLastName",
-                Name = "DemoUser",
-                PhoneNumber = "1234567890",
-                UserName = "demouser@microsoft.com",
-                ZipCode = "98052",
-                State = "WA",
-                Street = "15703 NE 61st Ct",
-                SecurityNumber = "535",
-                NormalizedEmail = "DEMOUSER@MICROSOFT.COM",
-                NormalizedUserName = "DEMOUSER@MICROSOFT.COM",
-                SecurityStamp = Guid.NewGuid().ToString("D"),
-            };
+                new ApplicationUser
+                {
+                    CardHolderName = "DemoUser",
+                    CardNumber = "4012888888881881",
+                    CardType = 1,
+                    City = "Redmond",
+                    Country = "U.S.",
+                    Email = "demouser@microsoft.com",
+                    Expiration = "12/20",
+                    Id = Guid.NewGuid().ToString(),
+                    LastName = "DemoLastName",
+                    Name = "DemoUser",
+                    PhoneNumber = "1234567890",
+                    UserName = "demouser@microsoft.com",
+                    ZipCode = "98052",
+                    State = "WA",
+                    Street = "15703 NE 61st Ct",
+                    SecurityNumber = "535",
+                    NormalizedEmail = "DEMOUSER@MICROSOFT.COM",
+                    NormalizedUserName = "DEMOUSER@MICROSOFT.COM",
+                    SecurityStamp = Guid.NewGuid().ToString("D")
+                };
 
             user.PasswordHash = _passwordHasher.HashPassword(user, "Pass@word1");
 
-            return new List<ApplicationUser>()
+            return new List<ApplicationUser>
             {
                 user
             };
         }
 
-        static string[] GetHeaders(string[] requiredHeaders, string csvfile)
+        private static string[] GetHeaders(string[] requiredHeaders, string csvfile)
         {
-            string[] csvheaders = File.ReadLines(csvfile).First().ToLowerInvariant().Split(',');
+            var csvheaders = File.ReadLines(csvfile).First().ToLowerInvariant().Split(',');
 
             if (csvheaders.Count() != requiredHeaders.Count())
-            {
-                throw new Exception($"requiredHeader count '{ requiredHeaders.Count()}' is different then read header '{csvheaders.Count()}'");
-            }
+                throw new Exception(
+                    $"requiredHeader count '{requiredHeaders.Count()}' is different then read header '{csvheaders.Count()}'");
 
             foreach (var requiredHeader in requiredHeaders)
-            {
                 if (!csvheaders.Contains(requiredHeader))
-                {
                     throw new Exception($"does not contain required header '{requiredHeader}'");
-                }
-            }
 
             return csvheaders;
         }
 
-        static void GetPreconfiguredImages(string contentRootPath, string webroot, ILogger logger)
+        private static void GetPreconfiguredImages(string contentRootPath, string webroot, ILogger logger)
         {
             try
             {
-                string imagesZipFile = Path.Combine(contentRootPath, "Setup", "images.zip");
+                var imagesZipFile = Path.Combine(contentRootPath, "Setup", "images.zip");
                 if (!File.Exists(imagesZipFile))
                 {
                     logger.LogError($" zip file '{imagesZipFile}' does not exists.");
                     return;
                 }
 
-                string imagePath = Path.Combine(webroot, "images");
-                string[] imageFiles = Directory.GetFiles(imagePath).Select(file => Path.GetFileName(file)).ToArray();
+                var imagePath = Path.Combine(webroot, "images");
+                var imageFiles = Directory.GetFiles(imagePath).Select(file => Path.GetFileName(file)).ToArray();
 
-                using (ZipArchive zip = ZipFile.Open(imagesZipFile, ZipArchiveMode.Read))
+                using (var zip = ZipFile.Open(imagesZipFile, ZipArchiveMode.Read))
                 {
-                    foreach (ZipArchiveEntry entry in zip.Entries)
-                    {
+                    foreach (var entry in zip.Entries)
                         if (imageFiles.Contains(entry.Name))
                         {
-                            string destinationFilename = Path.Combine(imagePath, entry.Name);
-                            if (File.Exists(destinationFilename))
-                            {
-                                File.Delete(destinationFilename);
-                            }
+                            var destinationFilename = Path.Combine(imagePath, entry.Name);
+                            if (File.Exists(destinationFilename)) File.Delete(destinationFilename);
                             entry.ExtractToFile(destinationFilename);
                         }
                         else
                         {
                             logger.LogWarning($"Skip file '{entry.Name}' in zipfile '{imagesZipFile}'");
                         }
-                    }
                 }
             }
             catch (Exception ex)
